requirement.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.16
django-bootstrap==0.2.4
django-crispy-forms==2.3
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.54.1
image==1.5.33
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.23.5
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Wand==0.6.13
zipp==3.20.2
